Step 1: move the meeseeks out of there into an "executor" package
